Synthe.cpp: use environemnt variables to find files.
The libray now looks up SYNTHE3_SEGFILE (resp. SYNTHE3_TABFILE)
to find Michel.seg (resp. Synthe.tab).

For each variable, if it is not defined the library tries to load the
corresponding file from the current directory.

The default values for both files will be made configurable in a later
commit.